Do not split pod names too eagerly
Statefulsets pod names are written <name>-<idx>, so if name contains
a `-` it would result in a mangled pod name in the logs, because of
the straight "eat last 2 components" rules.
Now we "detect" that it is a numerical index, and thus remove only
the last component (the statefulset index).